move references to chains
