+++ src/entries/popup/pages/home/Points/PointsOnboardingSheet.tsx
@@ -11,7 +11,7 @@ import {
   ValidatePointsSignatureMutation,
 } from '~/core/graphql/__generated__/metadata';
 import { i18n } from '~/core/languages';
-import { SUPPORTED_MAINNET_CHAINS } from '~/core/references';
+import { SUPPORTED_MAINNET_CHAINS } from '~/core/references/chains';
 import { useCurrentAddressStore } from '~/core/state';
 import { KeychainType } from '~/core/types/keychainTypes';
 import { formatNumber } from '~/core/utils/formatNumber';
