+++ src/entries/popup/pages/settings/rpcs.tsx
@@ -8,7 +8,7 @@ import {
   SUPPORTED_CHAINS,
   SUPPORTED_CHAIN_IDS,
   getDefaultRPC,
-} from '~/core/references';
+} from '~/core/references/chains';
 import { selectUserAssetsDictByChain } from '~/core/resources/_selectors/assets';
 import { useCustomNetworkAssets } from '~/core/resources/assets/customNetworkAssets';
 import {
@@ -98,8 +98,6 @@ export function SettingsNetworksRPCs() {
   const { developerToolsEnabled } = useDeveloperToolsEnabledStore();
   const { rainbowChains, setActiveRPC, removeCustomRPC } =
     useRainbowChainsStore();
-
-  console.log('-- rainbowChains', rainbowChains);
 
   const rainbowChain = rainbowChains[Number(chainId)];
